feat: add lnsfSafeSync

+++ src/fs/symlink.test.ts
@@ -1,7 +1,7 @@
 import { homedir } from 'node:os'
 import path, { basename, join } from 'node:path'
 import fse from 'fs-extra'
-import { lnsfSafe } from './symlink'
+import { lnsfSafe, lnsfSafeSync } from './symlink'
 
 describe('symlink', () => {
   const target = import.meta.filename
@@ -10,11 +10,19 @@ describe('symlink', () => {
   it('lnsfSafe', async () => {
     // remove _path first
     fse.removeSync(_path)
-
     // first create is ok
     await lnsfSafe(target, _path)
     // second overwrite is ok
     await lnsfSafe(target, _path)
+  })
+
+  it('lnsfSafeSync', () => {
+    // remove _path first
+    fse.removeSync(_path)
+    // first create is ok
+    lnsfSafeSync(target, _path)
+    // second overwrite is ok
+    lnsfSafeSync(target, _path)
   })
 
   it('lnsfSafe on a normal file', async () => {
@@ -30,8 +38,21 @@ describe('symlink', () => {
     // should overwrite
     await lnsfSafe(target, _path, { onExistingFile: 'delete' })
   })
+  it('lnsfSafeSync on a normal file', () => {
+    // remove _path
+    fse.removeSync(_path)
+    // nornal file
+    fse.outputFileSync(_path, 'hello world')
 
-  it('supports ~ in _path', async ({ skip }) => {
+    // should throw error
+    expect(() => lnsfSafeSync(target, _path)).toThrowError(/normal file/)
+    expect(() => lnsfSafeSync(target, _path, { onExistingFile: 'throw' })).toThrowError(/normal file/)
+
+    // should overwrite
+    lnsfSafeSync(target, _path, { onExistingFile: 'delete' })
+  })
+
+  it('supports ~ in _path: lnsfSafe', async ({ skip }) => {
     if (!process.env.CI) {
       return skip()
     }
@@ -39,6 +60,18 @@ describe('symlink', () => {
     const dir = '__tmp_needle-kit-tests__'
     const _pathPart = `${dir}/test.txt`
     await lnsfSafe(__filename, `~/${_pathPart}`)
+
+    // clean up
+    fse.removeSync(path.join(homedir(), dir))
+  })
+  it('supports ~ in _path: lnsfSafeSync', ({ skip }) => {
+    if (!process.env.CI) {
+      return skip()
+    }
+
+    const dir = '__tmp_needle-kit-tests__'
+    const _pathPart = `${dir}/test.txt`
+    lnsfSafeSync(__filename, `~/${_pathPart}`)
 
     // clean up
     fse.removeSync(path.join(homedir(), dir))
